fetch/metadata/generated/script-module-import-dynamic-sec-fetch-dest.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <script type="module">
  'use strict';

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, [], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'script');
        });
  }, '');
  </script>
</html>
